backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
paho-mqtt==1.6.1
confluent-kafka==2.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.18
alembic==1.11.1
python-dotenv==1.0.0
ortools==9.6.2534
geopy==2.4.0
